% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{adaptive_thin_pairs}
\alias{adaptive_thin_pairs}
\title{Adaptive Pairwise Thinning}
\usage{
adaptive_thin_pairs(pos, W_max = 1e+07, K_target = 1000)
}
\arguments{
\item{pos}{SNP physical positions (bp).}

\item{W_max}{Maximum LD distance.}

\item{K_target}{Target number of pairwise comparisons.}
}
\value{
List with thinning partitions and factor.
}
\description{
Determines thinning partitions to control the expected number of
pairwise LD comparisons per SNP.
}
\keyword{internal}
