% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{coef_ld_dec}
\alias{coef_ld_dec}
\title{Fit LD Decay Model}
\usage{
coef_ld_dec(dt_strata, q = 0.95, b)
}
\arguments{
\item{dt_strata}{LD edge data table with strata.}

\item{q}{Quantile for LD binning.}

\item{b}{Background LD.}
}
\value{
Named vector with parameters \code{c}, \code{a}, \code{d0}.
}
\description{
Fits the nonlinear decay model
\deqn{r^2(d) = b + \frac{c - b}{1 + a(d - d_0)}}
to binned LD quantiles.
}
