% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{coef_ld_dec_smooth}
\alias{coef_ld_dec_smooth}
\title{Fit LD Decay Model}
\usage{
coef_ld_dec_smooth(el_ld, q = 0.95, dist_unit = 5000, b = 0.05, min_bins = 5)
}
\arguments{
\item{el_ld}{LD edge data table.}

\item{q}{Quantile for LD binning.}

\item{dist_unit}{Distance bin size.}

\item{b}{Background LD.}
}
\value{
Named vector with parameters \code{c}, \code{a}, \code{d0}.
}
\description{
Fits the nonlinear decay model
\deqn{r^2(d) = b + \frac{c - b}{1 + a(d - d_0)}}
to binned LD quantiles.
}
\keyword{internal}
