% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_score.R
\name{combine_consistency}
\alias{combine_consistency}
\title{Add consistency scores to SNP map}
\usage{
combine_consistency(consistency_obj)
}
\arguments{
\item{consistency_obj}{Object of class "ld_consistency"}

\item{map}{Data.table with column \code{marker}}
}
\value{
Updated data.table
}
\description{
Add consistency scores to SNP map
}
