% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_summary.R
\name{compute_ld_summary}
\alias{compute_ld_summary}
\title{Compute LD-based summary statistics from histogram representation}
\usage{
compute_ld_summary(ld_structure, method = c("ld_int", "ld_w"), rho, cores = 1)
}
\arguments{
\item{ld_structure}{Object of class "ld_structure"}

\item{method}{"ld_int" (kernel integration) or "ld_w" (hard truncation)}

\item{rho}{distance from rho; d_window = rho / (a * (1 - rho))}

\item{cores}{Number of cores}

\item{eps}{Tail tolerance for LD_int}

\item{d_window}{Physical cutoff for ld_w (if NULL and method="ld_w",
it is derived from eps)}
}
\description{
Compute LD-based summary statistics from histogram representation
}
