% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eMLG_complexity_reduction.R
\name{default_cluster_colours}
\alias{default_cluster_colours}
\title{Default cluster / region colour palette}
\usage{
default_cluster_colours()
}
\value{
A character vector of colours.
}
\description{
A vector of ~40 visually distinct colours for colouring LD clusters or outlier
regions, one colour per cluster (recycle it if there are more regions than
colours). Used by the plotting helpers and the outlier-region vignette.
}
