% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_regions.R
\name{detect_or}
\alias{detect_or}
\title{Detect outlier regions from LD-scaled scan}
\usage{
detect_or(
  q_vals,
  ld_struct,
  decay_obj,
  sign_th,
  sign_if = c("less", "greater"),
  rho_d,
  rho_ld,
  l_min = 2,
  ret_table = FALSE
)
}
\description{
Detect outlier regions from LD-scaled scan
}
