% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{get_bg_ld}
\alias{get_bg_ld}
\title{Convert Relative LD Threshold to Distance}
\usage{
get_bg_ld(gds, idx = NULL, n_sub = 5000, q = 0.95)
}
\arguments{
\item{a}{Decay rate.}

\item{rho}{Relative LD threshold.}

\item{d0}{Offset parameter.}
}
\value{
Distance in base pairs.
}
\description{
Computes the physical distance corresponding to a relative LD
threshold \eqn{\rho}.
}
\keyword{internal}
