% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{ld_from_rho}
\alias{ld_from_rho}
\title{Convert Relative LD Threshold to Expected LD Value}
\usage{
ld_from_rho(b, c = 1, rho)
}
\arguments{
\item{b}{Background LD.}

\item{c}{Short-range LD (default = 1).}

\item{rho}{Relative LD threshold.}
}
\value{
Expected \eqn{r^2}.
}
\description{
Computes the expected \eqn{r^2} corresponding to a relative LD threshold \eqn{\rho}.
}
