% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDscn_pipeline.R
\name{LDscn_pipeline}
\alias{LDscn_pipeline}
\title{Run complete LD-aware genome scan pipeline}
\usage{
LDscn_pipeline(
  geno = NULL,
  map,
  q = 0.95,
  n_sub = 5000,
  window_size = 1e+06,
  step_size = 5e+05,
  dist_unit = 5000,
  K_target = 1000,
  n_win = 1000,
  prob_robust = 0.5,
  use = "robust",
  F_cols,
  q_cols = NULL,
  n_rho = 40,
  n_or_draws = 25,
  rho_w_lim = list(min = 0.5, max = 0.95),
  rho_d_lim = list(min = 0.9, max = 0.999),
  rho_ld_lim = list(min = 0.5, max = 0.999),
  alpha_lim = list(min = 1.31, max = 4),
  lmin_lim = list(min = 1, max = 10),
  cores = 1,
  max_rho = 0.99,
  verbose = TRUE
)
}
\arguments{
\item{geno}{Genotype matrix in 012-format. Expected genotypes are rounded to nearest integer for gds.}

\item{map}{SNP metadata; must contain Chr, Pos, marker and columns for the relevant test statistics.}

\item{F_cols}{Vector of column names for F-statistics .}

\item{q_cols}{Optional vector of q-values.}

\item{n_rho}{Number of LD-decay draws for robustness.}

\item{rho_w_lim}{Range of LD-decay quantiles for robustness.}

\item{rho_d_lim}{Range of LD-decay quantiles for robustness.}

\item{rho_ld_lim}{Range of LD-decay quantiles for robustness.'}

\item{cores}{N cores.}

\item{rho_w}{LD-decay quantile for single LD-scaled scan.}

\item{keep_draws}{keep draws.}

\item{keep_ld_str}{keep ld_structure.}

\item{keep_ld_decay}{keep decay.}

\item{detect_OR}{Logical; whether to detect ORs.}

\item{sign_th}{Significance threshold for OR detection.}

\item{sign_if}{"less" (q-values) or "greater" (C-score).}

\item{...}{Additional arguments passed to ld_rho_draws().}
}
\description{
Run complete LD-aware genome scan pipeline
}
