% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{plot_ldscn_manhattan}
\alias{plot_ldscn_manhattan}
\title{Plot Manhattan-style visualization for an LDscn object}
\usage{
plot_ldscn_manhattan(
  x,
  SNP_res,
  y_vars = c("C_mean"),
  y_labels = NULL,
  thresholds = NULL,
  compute_ld_w = FALSE,
  rho_w = 0.9,
  detect_OR = TRUE,
  sign_th = 0.05,
  sign_if = "greater",
  rho_d = 0.95,
  rho_ld = 0.95,
  l_min = 1,
  ncol = 1
)
}
\arguments{
\item{x}{An object returned by \code{LDscn_pipeline()}.}

\item{SNP_res}{An object returned by \code{LDscn_pipeline()}.}

\item{y_vars}{Character vector of variables to plot.}

\item{y_labels}{Optional labels for each panel.}

\item{thresholds}{Optional numeric vector of horizontal thresholds.}

\item{compute_ld_w}{Logical. Should LD-weights be computed?}

\item{rho_w}{Numeric. Quantile parameter for LD-weight computation.}

\item{detect_OR}{Logical. Should outlier regions be detected?}

\item{sign_th}{Significance threshold for OR detection.}

\item{sign_if}{Direction of significance ("less" or "greater").}

\item{rho_d}{LD-distance quantile for OR detection.}

\item{rho_ld}{LD-strength quantile for OR detection.}

\item{l_min}{Minimum OR size.}

\item{ncol}{Number of columns in patchwork layout.}
}
\value{
A \code{patchwork} object.
}
\description{
High-level Manhattan plotting wrapper for objects returned by
\code{LDscn_pipeline()}. This function:
}
\details{
\enumerate{
\item Adds consistency scores to the map.
\item Optionally computes LD-weights (\code{ld_w}).
\item Optionally detects outlier regions (ORs).
\item Generates one or multiple Manhattan panels.
}
}
