% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_rho_draws.R
\name{precalculate_ld_w}
\alias{precalculate_ld_w}
\title{Pre-calculates \code{ld_w}'s for \code{"ld_rho_draws"}}
\usage{
precalculate_ld_w(rho, ld_decay)
}
\arguments{
\item{rho}{A vector o rho values. Can be a single value.}

\item{ld_decay}{Object of class \code{"ld_decay"} returned by
\code{compute_LD_decay()}.}
}
\value{
matrix of \code{ld_w}'s for SNPs with each column corresponding to a different rho value
}
\description{
Pre-calculates \code{ld_w}'s for use with \code{"ld_rho_draws"}. Saves time, especially if edge lists are saved to file by \code{"compute_LD_decay"}
}
