% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{print.ld_decay}
\alias{print.ld_decay}
\title{Print an \code{ld_decay} object}
\usage{
\method{print}{ld_decay}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{ld_decay}.}

\item{digits}{Number of significant digits to display.}

\item{...}{Ignored (for S3 \code{print} consistency).}
}
\value{
Invisibly, \code{x}.
}
\description{
Print an \code{ld_decay} object
}
