% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{print.ld_structure}
\alias{print.ld_structure}
\title{Print Method for ld_structure Objects}
\usage{
\method{print}{ld_structure}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ld_structure"}.}

\item{...}{Additional arguments (unused).}
}
\description{
Displays a concise summary of the LD structure object.
}
