% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{coef_ld_dec}
\alias{coef_ld_dec}
\title{Fit LD Decay Model}
\usage{
coef_ld_dec(
  dt,
  q = 0.95,
  b,
  n_bins = 40,
  min_dist_quantile = 0.02,
  return_full = FALSE
)
}
\arguments{
\item{q}{Quantile for LD binning.}

\item{b}{Background LD.}

\item{el_ld}{LD edge data table.}

\item{dist_unit}{Distance bin size.}
}
\value{
Named vector with parameters \code{c}, \code{a}, \code{d0}.
}
\description{
Fits the nonlinear decay model
\deqn{r^2(d) = b + \frac{c - b}{1 + a(d - d_0)}}
to binned LD quantiles.
}
\keyword{internal}
