% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{compute_LD_decay}
\alias{compute_LD_decay}
\title{Compute Chromosome-wise LD Structure}
\usage{
compute_LD_decay(
  gds,
  el_data_folder = NULL,
  q = 0.95,
  n_sub_bg = 5000,
  n_win_decay = 20,
  overlap = 0.5,
  max_SNPs_decay = Inf,
  prob_robust = 0.95,
  max_pairs = 5000,
  n_strata = 20,
  keep_el = FALSE,
  slide = 1000,
  rho_targets = c(0.9, 0.95, 0.99),
  cores = 1
)
}
\arguments{
\item{gds}{GDS file handle containing genotype data.}

\item{el_data_folder}{Optional path to folder where edge-list of LD values are saved. If \code{"NULL"}, data not saved.}

\item{q}{Quantile used for LD-decay fitting.}

\item{n_sub_bg}{Number of SNPs used for background LD estimation.}

\item{n_win_decay}{Number of sliding windows per chromosome for decay fitting.}

\item{overlap}{Proportion of overlap between consecutive decay windows.}

\item{prob_robust}{Central proportion of windows retained for robust decay estimation.}

\item{keep_el}{Whether to save edge-lists directly. Useful for benchmarking smaller simulated data.}

\item{cores}{Number of CPU cores used for parallel computations.}

\item{keep_rho_hist}{Whether to save histograms for estimating ld_w values.}

\item{target_dist_bins_for_decay}{Number of distance bins for decay fitting.}

\item{k_max}{Slide window for LD-estimation.}

\item{n_rho_bins_hist}{Number of distance bins for saved LD histograms.}

\item{rho}{Maximum distance for pair-wise LD-values relative to decay rate.}

\item{r2_unit}{Bin width for r² values in histograms.}
}
\value{
An object of class \code{"ld_structure"} containing:
\describe{
\item{by_chr}{Per-chromosome LD structure, including histograms and decay parameters.}
\item{decay_sum}{Genome-wide table of robust decay parameters.}
\item{params}{List of parameters used for computation.}
}
}
\description{
Estimates linkage disequilibrium (LD) structure and LD-decay
parameters for each chromosome and (i)ntegrated LD (D)ecay (P)ersistence (I)ndex (iDPI) and optionally compresses and saved per-SNP LD
data for downstream LD-based summary statistics.
}
\details{
The workflow consists of:
\enumerate{
\item Estimation of background LD from inter-chromosomal pairs.
\item Sliding-window estimation of LD-decay parameters per chromosome.
\item Robust summarization of decay parameters.
\item Estimation of (i)ntegrated LD (D)ecay (P)ersistence (I)ndex (iDPI)
\item Optional saving of LD-data to files and histogram compression for downstream estimation of ld_w
}

LD decay is modeled as:
\deqn{r^2(d) = b + \frac{c - b}{1 + a d}}

where:
\itemize{
\item \eqn{a} is the decay rate,
\item \eqn{b} is background LD,
\item \eqn{c} is short-distance LD,
\item \eqn{d} is physical distance (bp).
}
}
