% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_score.R
\name{consistency_score}
\alias{consistency_score}
\title{Compute consistency score across rho_w and OR draws}
\usage{
consistency_score(or_dt)
}
\arguments{
\item{rho_draws_obj}{Object of class "ld_rho_draws".}
}
\value{
Object of class "ld_consistency".
}
\description{
Calculates SNP-level consistency across all rho_w and OR draws,
separately for each method.
}
