% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_regions.R
\name{detect_or}
\alias{detect_or}
\title{Detect outlier regions from LD-scaled scan}
\usage{
detect_or(
  gds,
  q_vals,
  ld_struct,
  sign_th = 0.05,
  sign_if = c("less", "greater"),
  use = c("robust", "median"),
  rho_d = 0.99,
  rho_ld = 0.99,
  l_min = 1,
  mode = c("per_method", "joint"),
  ret_table = FALSE
)
}
\description{
Detect outlier regions from LD-scaled scan
}
