% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{estimate_background_ld}
\alias{estimate_background_ld}
\title{Estimate LD Decay for a Chromosome}
\usage{
estimate_background_ld(gds, n_sub = 5000, q = 0.95)
}
\arguments{
\item{gds}{GDS file handle.}

\item{q}{Quantile for LD summary (default 0.95).}

\item{chr_idx}{SNP indices for the chromosome.}

\item{snp_pos}{SNP physical positions (bp).}

\item{b}{Background LD.}

\item{window_size}{Window size (bp).}

\item{step_size}{Step size (bp).}

\item{dist_unit}{Distance bin size (bp).}

\item{cores}{Number of CPU cores.}

\item{n_win}{Target number of SNPs per window.}
}
\value{
A \code{data.table} with window-wise decay parameters.
}
\description{
Fits LD-decay curves in sliding genomic windows for a single chromosome.
}
\details{
Decay is estimated using nonlinear least squares applied to binned
quantiles of \eqn{r^2}.
}
\keyword{internal}
