% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld_structure.R
\name{estimate_decay_chr}
\alias{estimate_decay_chr}
\title{Estimate LD Decay for a Chromosome}
\usage{
estimate_decay_chr(
  gds,
  chr_idx,
  snp_pos,
  b,
  window_size = 1e+06,
  step_size = 5e+05,
  q = 0.95,
  dist_unit = 5000,
  cores = 1,
  n_win = 1000
)
}
\arguments{
\item{gds}{GDS file handle.}

\item{chr_idx}{SNP indices for the chromosome.}

\item{snp_pos}{SNP physical positions (bp).}

\item{b}{Background LD.}

\item{window_size}{Window size (bp).}

\item{step_size}{Step size (bp).}

\item{q}{Quantile for LD summary (default 0.95).}

\item{dist_unit}{Distance bin size (bp).}

\item{cores}{Number of CPU cores.}

\item{n_win}{Target number of SNPs per window.}
}
\value{
A \code{data.table} with window-wise decay parameters.
}
\description{
Fits LD-decay curves in sliding genomic windows for a single chromosome.
}
\details{
Decay is estimated using nonlinear least squares applied to binned
quantiles of \eqn{r^2}.
}
\keyword{internal}
