% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_rho_draws.R
\name{ld_rho_draws}
\alias{ld_rho_draws}
\title{Perform multiple LD-window (rho_w) draws with OR detection}
\usage{
ld_rho_draws(
  ld_struct,
  decay_obj,
  SNP_ids,
  n_inds,
  F_vals,
  q_vals = NULL,
  n_rho = 25,
  n_or_draws = 25,
  rho_w_lim = list(min = 0.8, max = 0.99),
  rho_d_lim = list(min = 0.5, max = 0.999),
  rho_ld_lim = list(min = 0.9, max = 0.999),
  alpha_lim = list(min = 1.31, max = 4),
  lmin_lim = list(min = 1, max = 10),
  cores = 1,
  seed = NULL
)
}
\arguments{
\item{ld_struct}{Object of class "ld_structure".}

\item{decay_obj}{Object of class "ld_decay".}

\item{SNP_ids}{Vector of SNP IDs in correct order.}

\item{n_inds}{Number of individuals.}

\item{F_vals}{Matrix or data.frame of F statistics.}

\item{q_vals}{Matrix or data.frame of q values (optional).}

\item{n_rho}{Number of LD-window draws.}

\item{n_or_draws}{Number of OR parameter draws per rho_w.}

\item{rho_w_lim}{List(min,max) for rho_w.}

\item{rho_d_lim}{List(min,max) for rho_d.}

\item{rho_ld_lim}{List(min,max) for rho_ld.}

\item{alpha_lim}{List(min,max) for alpha exponent.}

\item{lmin_lim}{List(min,max) for minimum OR size.}

\item{seed}{Optional random seed.}
}
\value{
Object of class "ld_rho_draws".
}
\description{
Repeats LD-scaling across multiple LD-decay quantiles (rho_w),
and for each run performs multiple OR parameter draws.
}
