% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_scan.R
\name{ld_scan}
\alias{ld_scan}
\title{LD-scaled genome scan}
\usage{
ld_scan(ld_struct, decay_obj, F_vals, rho_w, n_inds, n_rep = 10, full = TRUE)
}
\arguments{
\item{ld_struct}{An object of class \code{"ld_structure"}.}

\item{decay_obj}{An object of class \code{"ld_decay"}.}

\item{F_vals}{Numeric vector or matrix of F-statistics (n_snp x n_method).}

\item{rho_w}{Numeric in (0,1). LD-decay quantile defining window size.}

\item{n_inds}{Number of individuals (for F null df2 = n_inds - 2).}

\item{n_rep}{Number of circular-shift permutations.}

\item{full}{Logical; if TRUE return QQ diagnostics.}
}
\value{
An object of class \code{"ld_scan"}.
}
\description{
Computes LD-scaled F statistics (F') for a given LD-decay object
and LD-structure, using a decay quantile \eqn{\rho_w}.
}
