% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDscn_pipeline.R
\name{LDscn_pipeline}
\alias{LDscn_pipeline}
\title{Run complete LD-aware genome scan pipeline}
\usage{
LDscn_pipeline(
  geno = NULL,
  map,
  F_cols,
  q_cols = NULL,
  slide_win_ld = 1000,
  n_rho = 40,
  n_or_draws = 25,
  rho_w_lim = list(min = 0.8, max = 0.99),
  rho_d_lim = list(min = 0.9, max = 0.999),
  rho_ld_lim = list(min = 0.5, max = 0.999),
  alpha_lim = list(min = 1.31, max = 4),
  lmin_lim = list(min = 1, max = 10),
  cores = 1,
  verbose = TRUE,
  keep_draws = FALSE,
  keep_ld_str = TRUE,
  keep_ld_decay = TRUE
)
}
\arguments{
\item{geno}{Genotype matrix in 012-format. Expected genotypes are rounded to nearest integer for gds.}

\item{map}{SNP metadata; must contain Chr, Pos, marker and columns for the relevant test statistics.}

\item{F_cols}{Vector of column names for F-statistics .}

\item{q_cols}{Optional vector of q-values.}

\item{n_rho}{Number of LD-decay draws for robustness.}

\item{rho_w_lim}{Range of LD-decay quantiles for robustness.}

\item{rho_d_lim}{Range of LD-decay quantiles for robustness.}

\item{rho_ld_lim}{Range of LD-decay quantiles for robustness.'}

\item{cores}{N cores.}

\item{keep_draws}{keep draws.}

\item{keep_ld_str}{keep ld_structure.}

\item{keep_ld_decay}{keep decay.}

\item{rho_w}{LD-decay quantile for single LD-scaled scan.}

\item{detect_OR}{Logical; whether to detect ORs.}

\item{sign_th}{Significance threshold for OR detection.}

\item{sign_if}{"less" (q-values) or "greater" (C-score).}

\item{...}{Additional arguments passed to ld_rho_draws().}
}
\description{
Run complete LD-aware genome scan pipeline
}
