% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDscnR-package.R
\docType{package}
\name{LDscnR-package}
\alias{LDscnR}
\alias{LDscnR-package}
\title{LDscnR: LD-aware detection of outlier regions in population genomic data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Petri Kemppainen \email{petri@genomeinsights.fi}

}
\keyword{internal}
