% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_regions.R
\name{or_draws}
\alias{or_draws}
\title{Draw ORs across parameter space (flat output)}
\usage{
or_draws(
  q_vals,
  ld_struct,
  decay_obj,
  n_draws = 25,
  rho_d_lim = list(min = 0.5, max = 0.999),
  rho_ld_lim = list(min = 0.9, max = 0.999),
  alpha_lim = list(min = 1.31, max = 4),
  lmin_lim = list(min = 1, max = 10)
)
}
\description{
Draw ORs across parameter space (flat output)
}
