% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_decay.R
\name{plot.ld_decay}
\alias{plot.ld_decay}
\title{Plot LD-decay curves}
\usage{
\method{plot}{ld_decay}(x, chr = NULL, max_dist = NULL, n_points = 500, ...)
}
\arguments{
\item{x}{An object of class \code{"ld_decay"}.}

\item{chr}{Optional chromosome to plot.}

\item{max_dist}{Maximum distance (bp) shown on x-axis.}

\item{n_points}{Number of points used to draw curve.}

\item{...}{Additional arguments.}
}
\description{
Visualizes fitted LD-decay curves based on per-chromosome
median parameters.
}
