% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_scan.R
\name{plot.ld_scan}
\alias{plot.ld_scan}
\title{#' Plot LD-scan results}
\usage{
\method{plot}{ld_scan}(x, method)
}
\arguments{
\item{x}{An object of class "ld_scan".}

\item{...}{Unused.}
}
\value{
A patchwork ggplot object.
}
\description{
Produces a two-panel figure:
\enumerate{
\item Quantile decomposition (LD distortion + selection excess)
\item Comparison of original, LD-weighted, and LD-scaled statistics
}
}
