% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_scan.R
\name{plot.ld_scan}
\alias{plot.ld_scan}
\title{Plot LD-scaled quantile transformation}
\usage{
\method{plot}{ld_scan}(x, method = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"ld_scan"}.}

\item{method}{Optional method name if multiple statistics were supplied.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Visualizes the quantile transformation underlying the LD-scaled
statistic, decomposing LD-induced distortion and
selection-consistent excess.
}
